+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/feedback/core/CarFeedbackSearchOptionsProvider.kt
@@ -1,0 +1,5 @@
+package com.mapbox.androidauto.car.feedback.core
+
+fun interface CarFeedbackSearchOptionsProvider {
+    fun searchOptions(): CarFeedbackSearchOptions
+}